+++ BTUtils/UITabBarController/UITabBarController+BTUtils.h
@@ -4,7 +4,7 @@
 //  Version 1.3
 //
 //  Created by Borut Tomazin on 8/30/2013.
-//  Copyright 2013 Borut Tomazin
+//  Copyright 2014 Borut Tomazin
 //
 //  Distributed under the permissive zlib License
 //  Get the latest version from here:
@@ -36,14 +36,12 @@
 
 /**
  Shows tabBar with/without animation.
- 
  @param animated
  */
 - (void)showTabBarAnimated:(BOOL)animated;
 
 /**
  Hides tabBar with/without animation.
- 
  @param animated
  */
 - (void)hideTabBarAnimated:(BOOL)animated;